requirements.txt
fastapi==0.104.1
pydantic==2.5.2
google-generativeai==0.5.2
